Reimplement for dealing with overlapped scroll view
