+++ Assets/Project/Prefabs/StageSelectScene/levelContents.prefab
@@ -325,7 +325,7 @@ GameObject:
   m_Component:
   - component: {fileID: 5194075750014507303}
   - component: {fileID: 7518556788165535580}
-  - component: {fileID: 6083101478836078306}
+  - component: {fileID: 8352029261218992239}
   m_Layer: 5
   m_Name: ScrollView
   m_TagString: Untagged
@@ -361,7 +361,7 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 9012854172227875165}
   m_CullTransparentMesh: 0
---- !u!114 &6083101478836078306
+--- !u!114 &8352029261218992239
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -370,7 +370,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 9012854172227875165}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: f4e0329b311fd4054b2acbc4ed5fcea2, type: 3}
+  m_Script: {fileID: 11500000, guid: ef8d90045570840c09e21c47df4c219f, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   m_Content: {fileID: 7759992107538557377}
